Create README.md
Today, with the rapid growth of mobile applications and merging with the concept of cloud processing, mobile cloud processing has been introduced as a potential technology for mobile services. This technology integrates cloud computing in the mobile environment, but it should include limitations such as battery life and processing power for mobile devices. To eliminate these constraints, it's a good idea to post components of a software to the cloud computing environment; previous research in this area focused on Single-Site Offloading. But with the advancement of technology today, it's easy to get into an environment in which there are several server deployments with different requirements, such as processing speed and bandwidth, and by dividing the optimal component of the software between these servers, it can reduce the amount of energy on the mobile. The proposed method in this study, using the ant colony algorithm, splits the implementation of software components between servers and reduces the runtime and, consequently, reduces the power consumption of the mobile device. It also prevents the overload phenomenon by reducing the running time by providing a method for balancing load between servers. A further method based on the multi-objective ant colony algorithm has been proposed, which is trying to solve the problem by considering both the energy constraints and the bar codes. The results show that the proposed methods have a better performance than the previous similar methods in reducing energy consumption and overhead phenomenon, and are more successful in the low number of servers and the limited computing resources.